+++ src/main/java/pl/futurecollars/invoicing/service/InvoiceService.java
@@ -15,11 +15,11 @@ public class InvoiceService {
     this.database = database;
   }
 
-  public int save(Invoice invoice) {
+  public Long save(Invoice invoice) {
     return database.save(invoice);
   }
 
-  public Optional<Invoice> getById(int id) {
+  public Optional<Invoice> getById(Long id) {
     return database.getById(id);
   }
 
@@ -27,11 +27,11 @@ public class InvoiceService {
     return database.getAll();
   }
 
-  public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+  public Optional<Invoice> update(Long id, Invoice updatedInvoice) {
     return database.update(id, updatedInvoice);
   }
 
-  public Optional<Invoice> delete(int id) {
+  public Optional<Invoice> delete(Long id) {
 
     return database.delete(id);
   }
